+++ R/zzz.R
@@ -9,7 +9,8 @@
         title = "R progress bar",
         label = "",
         nout = 100L,
-        min_time = 0)
+        min_time = 0,
+        use_lb = FALSE)
     optsx <- getOption("pboptions")
     if (!is.null(optsx)) {
         for (i in intersect(names(opts), names(optsx)))